LPS-59593 Only execute destructive actions if target and source JDBC connections are equals
This means that we want to backup the table on the same schema. On the other hand, if JDBC connections are different, it means that we are copying the table to the target connection, and we do not want to remove/drop data on the source schema.

+++ portal-impl/src/com/liferay/portal/upgrade/util/BaseUpgradeTableImpl.java
@@ -81,7 +81,7 @@ public abstract class BaseUpgradeTableImpl extends Table {
 		Connection connection = DataAccess.getUpgradeOptimizedConnection();
 
 		try {
-			updateTable(connection, connection);
+			updateTable(connection, connection, true);
 		}
 		finally {
 			DataAccess.cleanUp(connection);
@@ -89,7 +89,8 @@ public abstract class BaseUpgradeTableImpl extends Table {
 	}
 
 	public void updateTable(
-			Connection sourceConnection, Connection targetConnection)
+			Connection sourceConnection, Connection targetConnection,
+			boolean deleteFromSource)
 		throws Exception {
 
 		_calledUpdateTable = true;
@@ -101,7 +102,7 @@ public abstract class BaseUpgradeTableImpl extends Table {
 		try {
 			DB db = DBFactoryUtil.getDB();
 
-			if (Validator.isNotNull(tempFileName)) {
+			if (Validator.isNotNull(tempFileName) && deleteFromSource) {
 				String deleteSQL = getDeleteSQL();
 
 				db.runSQL(sourceConnection, deleteSQL);
@@ -110,7 +111,9 @@ public abstract class BaseUpgradeTableImpl extends Table {
 			String createSQL = getCreateSQL();
 
 			if (Validator.isNotNull(createSQL)) {
-				db.runSQL(sourceConnection, "drop table " + getTableName());
+				if (deleteFromSource) {
+					db.runSQL(sourceConnection, "drop table " + getTableName());
+				}
 
 				db.runSQL(targetConnection, createSQL);
 			}
